// frontend/src/app/api-gen/models/deployments-branches-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DeploymentEnvInfo } from '../models/deployment-env-info';
export interface DeploymentsBranchesResponse {
  cluster?: (string | null);
  environments: Array<DeploymentEnvInfo>;
  group_name: string;
  repo: string;
  service_name: string;
  task_definition_type?: (string | null);
}